refactor(img): rename Image → Img
Reverts the naming back to Img.
